backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
pymysql==1.1.0
cryptography==41.0.4
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
